+++ CHANGELOG.md
@@ -17,7 +17,7 @@ CHANGELOG
 #### Parsers
 - Shadowserver parser:
   * SSL FREAK: Remove optional column `device_serial` and add several new ones.
-- Spamhaus CERT parser: add support for `smtpauth` and `l_spamlink` (#1254).
+- Spamhaus CERT parser: add support for `smtpauth`, `l_spamlink`, `pop`, `imap`, `rdp` and `smb` (#1254).
 
 #### Experts
 - `bots.experts.filter`: Pre-compile regex at bot initialization.